requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
awsebcli==3.18.1
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.3.0
bitmex==0.2.2
bleach==3.1.5
blessed==1.17.5
botocore==1.15.49
cached-property==1.5.1
ccxt~=1.35.81
celery==4.4.2
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-celery-beat==2.0.0
django-environ==0.4.5
django-extensions==2.2.9
django-ipware==2.1.0
django-pandas==0.6.1
django-redis-cache==2.1.1
django-structlog==1.5.2
django-timezone-field==4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
future==0.16.0
h11==0.9.0
httptools==0.1.1
idna==2.7
idna-ssl==1.1.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pandas-datareader==0.8.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
pipdeptree==0.13.2
pprint==0.1
prettyjson==1.0.3
prometheus-client==0.7.1
psycopg2==2.8.5
ptyprocess==0.6.0
pycares==3.1.1
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-crontab==2.4.2
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
redis==3.5.1
requests==2.20.1
rfc3987==1.3.8
semantic-version==2.5.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.11.0
sqlparse==0.3.1
strict-rfc3339==0.7
structlog==20.1.0
swagger-spec-validator==2.5.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.24.3
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
wcwidth==0.1.9
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
whitenoise==5.0.1
yarl==1.1.0
zipp==3.1.0
django-picklefield~=3.0.1
ccxtpro~=0.4.4